+++ server/map.h
@@ -8,16 +8,16 @@
 
 struct Map {
 	char *data;
-	size_t width;
-	size_t height;
+	unsigned int width;
+	unsigned int height;
 	size_t size;
 };
 
-void map_write(int, char *, size_t, size_t);
+void map_write(int, char *, unsigned int, unsigned int);
 void map_free(struct Map *);
-void map_create(struct Map *, size_t, size_t);
+void map_create(struct Map *, unsigned int, unsigned int);
 void map_init_random(struct Map *, char *, size_t);
-int map_wrap(int, size_t);
+int map_wrap(int, unsigned int);
 char map_get(struct Map *, int, int);
 void map_set(struct Map *, int, int, char);
 int map_impassable(struct Map *, int, int);
